Be sure to define more services when creating a test code editor
